requirements.txt
celery
django 
redis 
requests
beautifulsoup4
djangorestframework
scraper